Split ShowPlanParser.cs into partial classes (#331)
* Add "Powered by Performance Studio" line on landing page


* Add Darling Data favicon to web app


* Add Open Graph and Twitter Card meta tags for social sharing

Uses the Darling Data barbell logo as hero image when shared on
social media. Also adds meta description for SEO.


* Clarify OG description: in-browser, nothing to install


* Fix Rule 3 severity: CouldNotGenerateValidParallelPlan is actionable

This reason means something in the query blocks parallelism (scalar UDFs,
table variable inserts, etc.) — that's worth a Warning, not Info.


* Expand Rule 3 to cover all NonParallelPlanReason values

Adds human-readable messages for all 25 known reasons. Severity:
- Warning: actionable reasons (UDFs, cursors, table variables, remote
  queries, trace flags, hints, DML OUTPUT, writeback variables)
- Info: passive/environmental (cost below threshold, edition limits,
  memory-optimized tables, upgrade mode, index build edge cases)


* Split ShowPlanParser.cs into partial classes

Move-only refactor; no behavior changes. ShowPlanParser.cs (1,844 lines)
split into 4 partials:

  RelOp      (789) - ParseRelOp (~760-line workhorse) +
                     GetOperatorElement + FindChildRelOps
  Warnings   (323) - ParseWarnings + ParseWarningsFromElement +
                     ParseMissingIndexes
  Costs      ( 25) - ComputeOperatorCosts + ComputeNodeCosts
  Helpers    ( 66) - CleanTempTableName, IsHexDigit, ScopedDescendants,
                     ParseColumnList, FormatColumnRef, ParseDouble/Long

Main file now 628 lines — class declaration, XML namespace, the public
Parse entry, ParseStatement{,AndChildren,Attributes},
ParseQueryPlanElements / ParseQueryPlanAsStatement.

PlanViewer.Web.csproj updated to link the new partial files.

Build clean: 0 errors, 0 warnings on PlanViewer.sln.


---------
